修复编译报错, 加 Vector2 类

+++ LearnDirectX12/Instance/InstanceApp.cpp
@@ -175,7 +175,7 @@ void InstanceApp::buildMaterial(dx12lib::CommonContextProxy pCommonCtx) {
 }
 
 void InstanceApp::buildPSO() {
-	UINT texArrayCount = std::min(kMaxTextureArraySize, _textures.size());
+	UINT texArrayCount = static_cast<UINT>(std::min(kMaxTextureArraySize, _textures.size()));
 
 	dx12lib::RootSignatureDescHelper rootDesc(d3d::getStaticSamplers());
 	rootDesc.resize(5);
